Add relacionamentos n<->n entre adv, processos e clientes

+++ database/migrations/2024_06_22_171643_create_clientes_pessoa_fis.php
@@ -24,8 +24,8 @@ return new class extends Migration
                 ->unique('uc_cliente_cpf');
             $table->integer('rg');
             $table->string('email')->nullable();
-            $table->integer('sexo');
-            $table->integer('estado_civil');
+            $table->smallInteger('sexo');
+            $table->smallInteger('estado_civil');
             $table->string('nome_pai', 60)->nullable();
             $table->string('nome_mae', 60)->nullable();
             $table->integer('naturalidade');
